@@ -1,3 +1,23 @@
+/*
+ * Copyright (c) 2023 Rylan
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ */
 package com.example.server.controller;
 
 import com.example.server.feign.CommonReturnType;
@@ -9,6 +29,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 
 @RestController
 public class TestController {
@@ -18,10 +39,11 @@ public class TestController {
 
     @GetMapping("/test")
     public CommonReturnType test() {
-        ArrayList<String> list = new ArrayList<>(Arrays.asList("西红柿", "玉米"));
-        System.out.println(feignClient.getBatch(list));
-        System.out.println(feignClient.getByName("西红柿")); //timeout
-        return feignClient.test(new Material(1001L, "material", "icon", "分类", "desc"));
+        var list = List.of("西红柿", "玉米");
+        System.out.println("getBatch: " + feignClient.getBatch(list));
+        System.out.println(feignClient.test(new Material(1001L, "material", "icon", "分类", "desc")));
+        System.out.println(feignClient.getRecipeByName("西红柿牛腩饭"));
+        return feignClient.getByName("西红柿");
     }
 
 }
