@@ -1,6 +1,25 @@
+/*
+ * Copyright (c) 2023 Rylan
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ */
 package cn.rylan.handler;
 
-import cn.rylan.http.RequestInterceptor;
 import cn.rylan.http.Http;
 import cn.rylan.model.MethodTemplate;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -20,15 +39,16 @@ public class RequestHandler {
     public Object http(String url, String methodType) throws JsonProcessingException {
         Class<?> returnType = methodTemplate.getReturnType();
         Http http = new Http();
-
-        if (methodType.equals("GET")) {
-            return http.get(url, returnType);
+        Object result = null;
+        switch (methodType) {
+            case "GET":
+                result = http.get(url, returnType);
+                break;
+            case "POST":
+                result = http.post(url, methodTemplate.getBody(), returnType);
+                break;
         }
-        if (methodType.equals("POST")) {
-            Object body = methodTemplate.getBody();
-            return http.post(url, body, returnType);
-        }
-        return null;
+        return result;
     }
 
 
