@@ -1,10 +1,30 @@
+/*
+ * Copyright (c) 2023 Rylan
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ */
 package cn.rylan.handler;
 
 import cn.rylan.model.MethodTemplate;
 import cn.rylan.model.ServiceInstance;
-import lombok.Data;
 import org.apache.commons.lang.StringUtils;
 
+import java.lang.ref.WeakReference;
 import java.lang.reflect.Method;
 import java.util.Map;
 
@@ -59,7 +79,8 @@ public class URLHandler {
     }
 
     private String replace(String uri, String target, String param) {
-        String s = new String("{" + target + "}");
+        String s = "{" + target + "}";
+        System.out.println(s.hashCode());
         return uri.replace(s, param);
     }
 
